refactor(licenses): share deriveCustomizeByLinkId to kill listLicenseLinks N+1
