+++ server/src/internal/licenses/actions/links/listLicenseLinks.ts
@@ -3,7 +3,7 @@ import { serializePlanLicense } from "../../licenseResponseUtils.js";
 import { getFullLicenseProduct } from "../../licenseUtils.js";
 import { licenseItemRepo } from "../../repos/licenseItemRepo.js";
 import { planLicenseRepo } from "../../repos/planLicenseRepo.js";
-import { deriveCustomizeFromItems } from "../customize/deriveLicenseCustomize.js";
+import { deriveCustomizeByLinkId } from "../customize/deriveLicenseCustomize.js";
 
 export const listLicenseLinks = async ({
 	ctx,
@@ -22,33 +22,22 @@ export const listLicenseLinks = async ({
 		parentInternalProductIds: [parentProduct.internal_id],
 	});
 
-	const itemRowsByLinkId = await licenseItemRepo.listByPlanLicenseIds({
+	const itemRows = await licenseItemRepo.listByPlanLicenseIds({
 		db: ctx.db,
 		planLicenseIds: rows.map(({ planLicense }) => planLicense.id),
 	});
-	return await Promise.all(
-		rows.map(async ({ planLicense, licensePlanId }) => {
-			const licenseProduct = await getFullLicenseProduct({
-				ctx,
-				idOrInternalId: planLicense.license_internal_product_id,
-			});
-			return serializePlanLicense({
-				planLicense,
-				parentPlanId: parentProduct.id,
-				licensePlanId,
-				customize: deriveCustomizeFromItems({
-					ctx,
-					licenseProduct,
-					itemRows: {
-						entitlements: itemRowsByLinkId.entitlements.filter(
-							(row) => row.plan_license_id === planLicense.id,
-						),
-						prices: itemRowsByLinkId.prices.filter(
-							(row) => row.plan_license_id === planLicense.id,
-						),
-					},
-				}),
-			});
+	const customizeByLinkId = await deriveCustomizeByLinkId({
+		ctx,
+		links: rows,
+		itemRows,
+	});
+
+	return rows.map(({ planLicense, licensePlanId }) =>
+		serializePlanLicense({
+			planLicense,
+			parentPlanId: parentProduct.id,
+			licensePlanId,
+			customize: customizeByLinkId.get(planLicense.id) ?? null,
 		}),
 	);
 };
